+++ crates/rendering/src/utils/window.rs
@@ -2,13 +2,13 @@ use getset::{CopyGetters, Getters, MutGetters};
 use glutin::{window::Fullscreen, PossiblyCurrent, WindowedContext};
 
 // Get the default width and height of the starting window
-pub const DEFAULT_WINDOW_SIZE: vek::Extent2<u16> = vek::Extent2::new(1280, 720);
+pub const DEFAULT_WINDOW_SIZE: vek::Extent2<u32> = vek::Extent2::new(1280, 720);
 
 // A window class to organize things
 #[derive(Getters, CopyGetters, MutGetters)]
 pub struct Window {
     #[getset(get_copy = "pub")]
-    pub(crate) dimensions: vek::Extent2<u16>,
+    pub(crate) dimensions: vek::Extent2<u32>,
     #[getset(get = "pub", get_mut = "pub")]
     pub(crate) context: WindowedContext<PossiblyCurrent>,
     #[getset(get_copy = "pub")]
